Remove axi_ad9234
The axi_ad9234 core is currently unused. The core is based on an older
version of the axi_ad9680 core but as slowly diverged and is currently
untested. So remove the core.

New projects that want to interface with the AD9234 should use the generic
JESD204 core (axi_adc_jesd204), which is up-to-date, tested and can support
the AD9234.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>